FOODAUTH-48 Implement the second part of requested PR changes
